Bits occur in PrivMsg, not USERNOTICE
